--- awx/ui_next/src/screens/Inventory/InventoryGroupForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './InventoryGroupForm';
