+++ awx/ui_next/src/screens/Inventory/shared/InventoryGroupForm.jsx
